+++ view/objetivo_profissional_lista.php
@@ -45,7 +45,7 @@ foreach($objetivo_profissional_json as $registro) {
 	echo $TR . $TD . $LABEL . 'Pretensão salarial: ' . $array_pretensao_salarial[$objetivo_profissional_model['obj_pro_pretensao_salarial']] . $LABEL_ . $TD_ . $TR_; 
 	echo $TR . $TD . $LABEL . 'Contrato: ' . $array_contrato[$objetivo_profissional_model['obj_pro_contrato']] . $LABEL_ . $TD_ . $TR_;  
 	echo $TR . $TD . '<a href="../view/objetivo_profissional.php?' . $str . '">Editar</a>' . $TD_ . $TR_; 
-	echo $TR . $TD . '<a href="../controller/objetivo_profissional.php?acao=excluir&' . $str . '">Excluir</a>' . $TD_ . $TR_; 
+	echo $TR . $TD . '<a href="../controller/objetivo_profissional.php?acao=excluir&' . $str . ' " onclick="return confirmar();">Excluir</a>' . $TD_ . $TR_; 
 	echo $TR . $TD . $LABEL . '&nbsp;' . $LABEL_ . $TD_ . $TR_; 	
 }
